requirements.txt
uvicorn
streamlit
langchain
langchain-community
langsmith
python-dotenv
langchain-google-genai
google-generativeai
langchain-pinecone
fastapi
motor
streamlit-feedback
pydantic-settings
python-dotenv
langkit[all]

